Refactor pmd_common functions to static methods
